refactor: ViewController 통합
