+++ Calculator/Calculator/Source/Model/DoubleExtension.swift
@@ -2,7 +2,7 @@
 //  DoubleExtension.swift
 //  Calculator
 //
-//  Created by 송종환 on 2023/02/07.
+//  Created by Harry, Goat on 2023/02/07.
 //
 
 import Foundation
